Fixed semantic bug where log_idx = 5 could over-write log_idx = 6 but the recovery mechanism kept the intended order intact because of how the bytes were flushed to the disk but in any case now the order in which bytes are in the wal and the order in which the log index represent writes applied now completely match

+++ engine/src/headers/globals.h
@@ -20,9 +20,9 @@ struct write_cmd {
 };
 
 struct raft_stats {
-    uint8_t  forwarding_node{0};
-    uint8_t  leader_node{0};
-    uint64_t time_leader_received{0};
+    int  forwarding_node{0};
+    int  leader_node{0};
+    int time_leader_received{0};
 };
 
 struct entry {
